Fix logic preventing RAs from having duty with themselves
This implements checking that the primary RA on the last day of a
segment is not the only RA left in the secondary RA pool. If that is the
case, the RA is swapped into a random secondary duty in the same
segment, and the newly displaced RA which previously had that duty is
assigned to secondary on the last day in that segment.

Fixes #1